Remove `nanobench`
