+++ package.json
@@ -46,7 +46,6 @@
   "devDependencies": {
     "brittle": "^3.0.0",
     "is-ci": "^3.0.1",
-    "nanobench": "^2.1.1",
     "prebuildify": "^5.0.0",
     "standard": "^16.0.4",
     "tiny-byte-size": "^1.0.1"
